producer/requirements.txt
Faker==11.3.0
kafka-python==2.0.2
